requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
streamlit
click
pyyaml
networkx
sqlparse
python-dotenv